xen: credit2: consider soft-affinity in get_fallback_cpu()
so that, if the trylock-s fail in csched2_cpu_pick(),
the cpu we fall back on respects the vcpu's soft-affinity,
if one is present.

This is done by implementing the usual two steps soft/hard
affinity balancing loop. Inside each step, we give
preference to (in this order) the last processor where
the vcpu has been executing (for cache betterness), or a
processor from the runqueue the vcpu is assigned to (to
try avoiding migration).

Signed-off-by: Dario Faggioli <dario.faggioli@citrix.com>
Signed-off-by: Justin T. Weaver <jtweaver@hawaii.edu>
---
Cc: George Dunlap <george.dunlap@citrix.com>
Cc: Anshul Makkar <anshul.makkar@citrix.com>
Cc: David Vrabel <david.vrabel@citrix.com>