xen: credit2: the private scheduler lock can be an rwlock.
the private scheduler lock protects data that only really
changes either at init-time, during cpupools manipulation
and when changing domain weights. In all other cases
(namely, load balancing, reading weights and status dumping)
information is only read.

It is therefore possible to make it a read/write lock. This
is particularly good for the scalability of the load balancing
code, as we does not require full serialization with the
rest of the scheduler any longer.

A couple of ASSERT()-s are also added/updated, as well as
all the relevant comments.

While doing this, reorg and reword the high level descrition
of how locking works at the top of the file.

Signed-off-by: Dario Faggioli <dario.faggioli@citrix.com>
---
Cc: George Dunlap <george.dunalp@citrix.com>
Cc: Anshul Makkar <anshul.makkar@citrix.com>
Cc: David Vrabel <david.vrable@citrix.com>