xen: credit2: read NOW() with the proper runq lock held
Yet another situation very similar to 779511f4bf5ae
("sched: avoid races on time values read from NOW()").

In fact, when more than one runqueue is involved, we need
to make sure that the following does not happen:
 1. take the lock of 1st runq
 2. now = NOW()
 3. take the lock of 2nd runq
 4. use now

as, if we have to wait at step 3, the value in now may
be stale when we get to use it at step 4.

While there, fix the style of a label.

Signed-off-by: Dario Faggioli <dario.faggioli@citrix.com>
---
Cc: George Dunlap <george.dunlap@citrix.com>
Cc: Anshul Makkar <anshul.makkar@citrix.com>
Cc: David Vrabel <david.vrabel@citrix.com>